feat: DEV-2887: move Comment up to top level of stores
